convert autopickup range, speed, range, melee_range, threat_range to tile based units

+++ src/MapCollision.h
@@ -71,7 +71,7 @@ public:
 	bool is_wall(int x, int y) const;
 
 	bool is_valid_tile(int x, int y, MOVEMENTTYPE movement_type, bool is_hero) const;
-	bool is_valid_position(int x, int y, MOVEMENTTYPE movement_type, bool is_hero) const;
+	bool is_valid_position(float x, float y, MOVEMENTTYPE movement_type, bool is_hero) const;
 
 	int is_one_step_around(float x, float y, int xidr, int ydir);
